ctai_backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==41.0.7
